Strict quality control for nodes

+++ parser/asts/expr/add_expr.cpp
@@ -7,7 +7,7 @@ using namespace ast;
 
 std::string AddExprNode::to_string() {
     std::string ret = "AddExprNode: {[" + head.to_string() + ", ";
-    for(int i = 0; i < ops.size(); i ++) {
+    for(size_t i = 0; i < ops.size(); i ++) {
         ret += ops[i].to_string() + ", ";
         ret += factors[i].to_string() + ", ";
     }
